docs: Add README.build file.
This file is generated from the output of `make help`. The file
is extracted so that we have access to build system documentation
without executing anything.

README.build should not be changed directly, instead the help text
in Makefile should be edited and then `make README.build` would
rebuild the file. That way the help text is closer to Makefile
rules and more dynamic.

Another plus to having README.build file is that OSCam users now
can be pointed directly to this file in OSCam SVN and web site.

git-svn-id: https://svn.streamboard.tv/oscam/trunk@7416 4b0bc96b-bc66-0410-9d44-ebda105a78c1