Make get_cardsystem_desc_by_caid() function public and use it also in dvbapi/ecminfo
The function get_cardsystem_desc_by_caid() is used by CCcam share to print nice
cardsystem description in the webif.

This commit moves the above function from CCcam specific to oscam-simples.c.
This way it is also used in the dvbapi/ecminfo to print the cardsystem name
(eg. when displaing the ECM information in the skin data).


git-svn-id: https://svn.streamboard.tv/oscam/trunk@10653 4b0bc96b-bc66-0410-9d44-ebda105a78c1